requirements.txt
# app
PyYAML
bs4
fastapi
# git+git://github.com/arount/pytea#egg=pytea
git+git://github.com/linearcombination/USFM-Tools.git@develop#egg=usfm_tools
## For local dev, install via: pip install /path/to/USFM-Tools
git+git://github.com/unfoldingWord-dev/uw_tools.git#egg=uw_tools
gogs-client
icontract
jsonpath-rw
jsonpath-rw-ext
logdecorator
Markdown
# markdown
# markdown2
openpyxl
parse
pathlib
pydantic
pyparsing==2.4.5
requests
uvicorn

# tests
mypy
flake8
wemake-python-styleguide
flake8-mypy
pytest
pytest-icdiff
pylint
tenacity
